+++ src/Indice.Features.Messages.Core/Data/Mappings/DbMessageEventMap.cs
@@ -25,7 +25,7 @@ public class DbMessageEventMap : IEntityTypeConfiguration<DbMessageEvent>
         // Configure properties.
         builder.Property(x => x.Type).HasMaxLength(TextSizePresets.S64);
         builder.Property(x => x.Channel).HasMaxLength(TextSizePresets.S64);
-        builder.Property(x => x.Receiver).HasMaxLength(TextSizePresets.M128);
+        builder.Property(x => x.Recipient).HasMaxLength(TextSizePresets.M128);
         
         // indexes
         builder.HasIndex(x => x.Type);